backend/requirements.txt
Flask
flask-Cors
requests
youtube-transcript-api
python-dotenv
openai